+++ includes/traits/trait-stancer-refunds.php
@@ -30,30 +30,48 @@ trait WC_Stancer_Refunds_Traits {
 	 *
 	 * @param WC_Order $order the WooCommerce order to be refunded.
 	 * @return boolean
+	 *
+	 * @throws WC_Stancer_Exception Throw an exception if the config is misconfigured.
 	 */
 	public function can_refund_order( $order ) {
 		if ( $order->get_payment_method( 'view' ) !== $this->id ) {
 			return false;
 		}
-		if ( ! $order->payment_complete() ) {
+
+		if ( ! in_array( $order->get_status( 'view' ), [ 'completed', 'processing' ], true ) ) {
 			return false;
 		}
+
 		if ( $order->get_total() <= $order->get_total_refunded() ) {
 			return false;
 		}
+
 		$transaction_id = $order->get_transaction_id() ?? null;
+
 		if ( ! $transaction_id ) {
-			$api_payment = WC_Stancer_Payment::find( $order );
-			$transaction_id = $api_payment->payment_id;
+			$stancer_payment = WC_Stancer_Payment::find( $order );
+			$transaction_id = $stancer_payment->payment_id;
 		}
-		$api_payment = new Stancer\Payment( $transaction_id );
 
 		$status = [
 			Stancer\Payment\Status::TO_CAPTURE,
 			Stancer\Payment\Status::CAPTURED,
-			Stancer\Payment\Status::CAPTURED,
 		];
-		return in_array( $api_payment->getStatus(), $status, true );
+
+		try {
+			// Don't know why, but WC does not find the settings if did not do it myself.
+			$settings = get_option( 'woocommerce_stancer_settings' );
+			$wc_config = new WC_Stancer_Config( $settings );
+
+			if ( $wc_config->is_not_configured() ) {
+				throw new WC_Stancer_Exception( 'Your api is not properly configured' );
+			}
+			$api_payment = new Stancer\Payment( $transaction_id );
+
+			return in_array( $api_payment->get_status(), $status, true );
+		} catch ( Exception $e ) {
+			return false;
+		}
 	}
 
 	/**
@@ -71,14 +89,19 @@ trait WC_Stancer_Refunds_Traits {
 	 */
 	public function process_refund( $order_id, $amount = null, $reason = '' ): bool {
 		$wc_order = wc_get_order( $order_id );
+
 		if ( $this->api_config->is_not_configured() ) {
-			WC()->session->set( 'stancer_error_payment', __( 'The module is not correctly configured.', 'stancer' ) );
-			throw new Exception( esc_html( __( 'The module is not correctly configured', 'stancer' ) ) );
+			$message = __( 'The module is not correctly configured.', 'stancer' );
+
+			WC()->session->set( 'stancer_error_payment', $message );
+
+			throw new Exception( esc_html( $message ) );
 		}
 
 		$stancer_payment = $this->api->send_refund( $wc_order, $amount ? (int) ( $amount * 100 ) : null );
 		$refundable = $stancer_payment->getRefundableAmount();
 		$currency = $stancer_payment->currency;
+
 		if ( 0 !== $refundable ) {
 			$text = sprintf(
 				// translators: "%1$.2f": Amount refunded. "%2$s": Currency. "%3$.2f": Total amount after all refunds.