+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.2.1] - 2024-03-22
 
 ### Fixed
+- Fix some locales
 - Fix scoped dependencies on PHP 7.4