+++ CHANGELOG.md
@@ -4,10 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [Unreleased]
 
 ## Fixed
--Fix a bug that changed order statuses. (CMS-117)
+- Fix a bug that changed order statuses. (CMS-117)
+
 
 ## [1.2.1] - 2024-03-22