+++ CHANGELOG.md
@@ -4,9 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.2] - 2024-04-02
+
+### Fixed
+- Fix config bug (CMS-116)
+- Fix a bug that changed order statuses (CMS-117)
+
+
 ## [1.2.1] - 2024-03-22
 
 ### Fixed
+- Fix some locales
 - Fix scoped dependencies on PHP 7.4